[camera_avfoundation] support triple and double composite cameras on ios
This CL adds support for the dual and triple cameras that are
standard on modern iOS devices. Discovery will now include these
cameras, along with the others that were already there. The list
will be ordered from highest quality to lowest quality, ensuring
that `(await availableCameras()).first()` will generally be the
best default to choose -- the one that most accurately emulates
the camera functionality the user is accustomed to seeing in the
built-in camera app.